requirements.txt
django==1.2.5
mysql-python==1.2.3
ipython==0.10.1
django-picklefield==0.1.9
south==0.7.3
https://github.com/iplatform/pyFaceGraph/tarball/master